% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Distconvert.R
\name{DistConvert}
\alias{DistConvert}
\title{Distance Conversion}
\usage{
DistConvert(DistIn, DistOut, dist1)
}
\arguments{
\item{DistIn}{Choose unit of measurement for input, in (inches), ft (feet), yd (yards), mi (miles), mm (millimeters), cm (centimeters), m (meters), km (kilometers)}

\item{DistOut}{Choose unit of measurement for output, in (inches), ft (feet), yd (yards), mi (miles), mm (millimeters), cm (centimeters), m (meters), km (kilometers)}

\item{dist1}{Numeric value for input distance}
}
\value{
numeric value in the measurement chosen by DistOut
}
\description{
Function allows users to select an input and output unit of Distance, the current supported units are inches(in), feet (ft), yards (yd), miles (mi), millimeters (mm), centimeters (cm), meters (m), and kilometers (km)
}
\examples{
DistConvert("km","mi",100)

}
