% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Veloconvert.R
\name{VeloConvert}
\alias{VeloConvert}
\title{Velocity Conversion}
\usage{
VeloConvert(veloIn, veloOut, velo)
}
\arguments{
\item{veloIn}{Choose unit of measurement for input, mph (miles per hour), kph (kilometers per hour), m/s (meters per second)}

\item{veloOut}{Choose unit of measurement for output, mph (miles per hour), kph (kilometers per hour), m/s (meters per second)}

\item{velo}{Numeric value for input velocity}
}
\value{
numeric value in the measurement chosen by veloout
}
\description{
Function allows users to select an input and output unit of velocity, the current supported units are miles per hour (mph), kilometers per hour (kph), and meters per second (m/s)
}
\examples{
VeloConvert("mph","kph",60)

}
